android: Fix unlink error and add logging
Add logging to help debug a case where a benchmark without the build
server running would fail occasionally when the previous benchmark had
the build server on.

Bug: 383150489
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6089016
Commit-Queue: Andrew Grieve <agrieve@chromium.org>
Auto-Submit: Peter Wen <wnwen@chromium.org>
Commit-Queue: Peter Wen <wnwen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1394898}

+++ build/android/gyp/lint.py
@@ -389,7 +389,8 @@ def _RunLint(custom_lint_jar_path,
       shutil.rmtree(aar_root_dir, ignore_errors=True)
       shutil.rmtree(resource_root_dir, ignore_errors=True)
       shutil.rmtree(srcjar_root_dir, ignore_errors=True)
-      os.unlink(project_xml_path)
+      if os.path.exists(project_xml_path):
+        os.unlink(project_xml_path)
       shutil.rmtree(partials_dir, ignore_errors=True)
 
     if failed:
